MAINT move NotFittedError to sklearn.base
