+++ sklearn/cluster/birch.py
@@ -10,11 +10,12 @@ from scipy import sparse
 from math import sqrt
 
 from ..metrics.pairwise import euclidean_distances
-from ..base import TransformerMixin, ClusterMixin, BaseEstimator
+from ..base import (TransformerMixin, ClusterMixin, BaseEstimator,
+                    NotFittedError)
 from ..externals.six.moves import xrange
 from ..utils import check_array
 from ..utils.extmath import row_norms, safe_sparse_dot
-from ..utils.validation import NotFittedError, check_is_fitted
+from ..utils.validation import check_is_fitted
 from .hierarchical import AgglomerativeClustering
