+++ sklearn/random_projection.py
@@ -35,13 +35,13 @@ import numpy as np
 from numpy.testing import assert_equal
 import scipy.sparse as sp
 
-from .base import BaseEstimator, TransformerMixin
+from .base import BaseEstimator, TransformerMixin, NotFittedError
 from .externals import six
 from .externals.six.moves import xrange
 from .utils import check_random_state
 from .utils.extmath import safe_sparse_dot
 from .utils.random import sample_without_replacement
-from .utils.validation import check_array, NotFittedError
+from .utils.validation import check_array
 from .utils import DataDimensionalityWarning
